+++ Source/Core/APIRequest.swift
@@ -96,36 +96,39 @@ open class APIRequest<Model, ErrorModel>: BaseRequest<Model,ErrorModel> {
         guard let manager = tronDelegate?.manager else {
             fatalError("Manager cannot be nil while performing APIRequest")
         }
+        allPlugins.forEach { $0.willSendRequest(self) }
         guard let request = alamofireRequest(from: manager) as? DataRequest else {
             fatalError("Failed to receive DataRequest")
         }
+        allPlugins.forEach { $0.willSendAlamofireRequest(request, formedFrom: self) }
         if !tronDelegate!.manager.startRequestsImmediately {
             request.resume()
         }
-        // Notify plugins about new network request
-        let allPlugins = plugins + (tronDelegate?.plugins ?? [])
-        allPlugins.forEach {
-            $0.willSendRequest(request.request)
-        }
-        return request.validate().response(queue: resultDeliveryQueue,responseSerializer: dataResponseSerializer(notifyingPlugins: allPlugins), completionHandler: completion)
+        allPlugins.forEach { $0.didSendAlamofireRequest(request, formedFrom: self) }
+        
+        return request.validate().response(queue: resultDeliveryQueue,responseSerializer: dataResponseSerializer(with: request), completionHandler: completion)
     }
     
-    internal func dataResponseSerializer(notifyingPlugins plugins: [Plugin]) -> Alamofire.DataResponseSerializer<Model> {
-        return DataResponseSerializer<Model> { urlRequest, response, data, error in
+    internal func dataResponseSerializer(with request: Request) -> Alamofire.DataResponseSerializer<Model> {
+        return DataResponseSerializer<Model> { [weak self, allPlugins] urlRequest, response, data, error in
+            guard let weakSelf = self else {
+                debugPrint("Request deallocated before calling DataResponseSerializer")
+                return .failure(TRONError.requestDeallocated)
+            }
             DispatchQueue.main.async(execute: {
-                plugins.forEach {
-                    $0.requestDidReceiveResponse(urlRequest, response,data,error)
+                allPlugins.forEach {
+                    $0.didReceiveResponse(response: (urlRequest, response, data, error), forRequest: request, formedFrom: weakSelf)
                 }
             })
             if let error = error {
-                return .failure(self.errorParser(nil, urlRequest, response, data, error))
+                return .failure(weakSelf.errorParser(nil, urlRequest, response, data, error))
             }
             
-            let result = self.responseParser(urlRequest, response, data, error)
+            let result = weakSelf.responseParser(urlRequest, response, data, error)
             if let model = result.value {
                 return .success(model)
             } else {
-                return .failure(self.errorParser(result, urlRequest, response, data, error))
+                return .failure(weakSelf.errorParser(result, urlRequest, response, data, error))
             }
         }
     }