+++ Source/Core/DownloadAPIRequest.swift
@@ -101,35 +101,37 @@ open class DownloadAPIRequest<Model, ErrorModel>: BaseRequest<Model,ErrorModel> 
         guard let manager = tronDelegate?.manager else {
             fatalError("Manager cannot be nil while performing APIRequest")
         }
+        allPlugins.forEach { $0.willSendRequest(self) }
         guard let request = alamofireRequest(from: manager) as? DownloadRequest else {
             fatalError("Failed to receive DataRequest")
         }
+        allPlugins.forEach { $0.willSendAlamofireRequest(request, formedFrom: self) }
         if !tronDelegate!.manager.startRequestsImmediately {
             request.resume()
         }
-        // Notify plugins about new network request
-        let allPlugins = plugins + (tronDelegate?.plugins ?? [])
-        allPlugins.forEach {
-            $0.willSendRequest(request.request)
-        }
-        return request.validate().response(queue: resultDeliveryQueue, responseSerializer: downloadResponseSerializer(notifyingPlugins: allPlugins), completionHandler: completion)
+        allPlugins.forEach { $0.didSendAlamofireRequest(request, formedFrom: self)}
+        return request.validate().response(queue: resultDeliveryQueue, responseSerializer: downloadResponseSerializer(with: request), completionHandler: completion)
     }
     
-    internal func downloadResponseSerializer(notifyingPlugins plugins: [Plugin]) -> DownloadResponseSerializer<Model> {
-        return DownloadResponseSerializer<Model> { urlRequest, response, url, error in
+    internal func downloadResponseSerializer(with request: DownloadRequest) -> DownloadResponseSerializer<Model> {
+        return DownloadResponseSerializer<Model> { [weak self, allPlugins] urlRequest, response, url, error in
+            guard let weakSelf = self else {
+                debugPrint("Request deallocated before calling DownloadResponseSerializer")
+                return .failure(TRONError.requestDeallocated)
+            }
             DispatchQueue.main.async(execute: {
-                plugins.forEach {
-                    $0.requestDidReceiveResponse((urlRequest, response, nil,error))
+                allPlugins.forEach {
+                    $0.didReceiveResponse(response: (urlRequest, response,nil,error), forRequest: request, formedFrom: weakSelf)
                 }
             })
             if let error = error {
-                return .failure(self.errorParser(nil, urlRequest, response, url, error))
+                return .failure(weakSelf.errorParser(nil, urlRequest, response, url, error))
             }
-            let result = self.responseParser(urlRequest, response, url, error)
+            let result = weakSelf.responseParser(urlRequest, response, url, error)
             if let model = result.value {
                 return .success(model)
             } else {
-                return .failure(self.errorParser(result, urlRequest, response, url, error))
+                return .failure(weakSelf.errorParser(result, urlRequest, response, url, error))
             }
         }
     }