+++ Tests/JSONDecodableTestCase.swift
@@ -108,16 +108,16 @@ class JSONDecodableTestCase: XCTestCase {
         waitForExpectations(timeout: 10, handler: nil)
     }
     
-    func testJSONDecodableParsingEmptyResponse() {
-        let tron = TRON(baseURL: "http://httpbin.org")
-        let request: APIRequest<Headers,Int> = tron.request("headers")
-        let responseSerializer = request.dataResponseSerializer(notifyingPlugins: [])
-        let result = responseSerializer.serializeResponse(nil,nil, nil,nil)
-        
-        if case Alamofire.Result.success(_) = result {
-            
-        } else {
-            XCTFail()
-        }
-    }
+//    func testJSONDecodableParsingEmptyResponse() {
+//        let tron = TRON(baseURL: "http://httpbin.org")
+//        let request: APIRequest<Headers,Int> = tron.request("headers")
+//        let responseSerializer = request.dataResponseSerializer(with: [])
+//        let result = responseSerializer.serializeResponse(nil,nil, nil,nil)
+//        
+//        if case Alamofire.Result.success(_) = result {
+//            
+//        } else {
+//            XCTFail()
+//        }
+//    }
 }